requirements.txt
tensorflow==2.9.1
numpy==1.23.1
opencv-python==4.5.3.56
fastapi==0.68.1
uvicorn==0.15.0
pillow==9.2.0
streamlit==1.12.0
requests==2.28.1
